requirements.txt
dash==2.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash_bootstrap_components==1.1.0
pandas==1.4.2
plotly==5.8.2
python==3.9.12
psycopg2-binary==2.8.6
textblob==0.17.1
gunicorn==20.1.0
tweepy==3.10.0
nltk==3.7